# decision-agent/agent-backend/agent-factory/cmd/openapi-docs/generated/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agent Factory 智能体工厂 API 文档",
        "title": "Agent Factory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:30777",
    "basePath": "/api/agent-factory",
    "paths": {
...
        "/v1/app/{app_key}/conversation": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定应用的会话列表，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对话管理"
                ],
                "summary": "获取对话列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 Key",
                        "name": "app_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量，默认10",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的会话会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对话管理"
                ],
                "summary": "初始化对话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 Key",
                        "name": "app_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "初始化请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/conversationreq.InitReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除所有对话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对话管理"
                ],
                "summary": "删除所有对话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_key",
                        "name": "app_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "所有对话删除成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/v1/app/{app_key}/conversation/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据会话 ID 获取会话详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对话管理"
                ],
                "summary": "获取对话详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用 Key",
                        "name": "app_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "会话不存在",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "编辑对话(修改标题)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对话管理"
                ],
                "summary": "编辑对话(修改标题)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_key",
                        "name": "app_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除对话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对话管理"
                ],
                "summary": "删除对话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_key",
                        "name": "app_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "对话不存在",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "conversationreq.InitReq": {
            "type": "object",
            "properties": {
                "agent_id": {
                    "description": "agent id",
                    "type": "string"
                },
                "agent_version": {
                    "description": "agent version",
                    "type": "string"
                },
                "executor_version": {
                    "description": "executor version v1 或 v2 默认v2",
                    "type": "string"
                },
                "temparea_id": {
                    "description": "临时区域ID",
                    "type": "string"
                },
                "title": {
                    "description": "conversation 标题",
                    "type": "string"
                },
                "visitor_type": {
                    "description": "访客类型  user / app",
                    "type": "string"
                }
            }
        },
...
        "swagger.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "details": {
                    "type": "string",
                    "example": "详细错误信息"
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer {token} - 需要有效的 OAuth 访问令牌",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}